Optimize Content-Length: 0 handling
Create and use a pre-encoded HttpFields.CONTENT_LENGTH_0 constant
Used the constant in more places
renamed the `putLongField` methods of HttpFields.Mutable to just `put`

+++ jetty-core/jetty-http/src/main/java/org/eclipse/jetty/http/PreEncodedHttpField.java
@@ -126,7 +126,7 @@ public class PreEncodedHttpField extends HttpField
     {
         if (_longValue == Long.MIN_VALUE)
             return super.getIntValue();
-        return (int)_longValue;
+        return Math.toIntExact(_longValue);
     }
 
     @Override
